Number and Alphabet Guessing Game
The Guessing Game Program’s Concept is asking the user to guess a randomly generated number or alphabet within a given range, depending on the level chosen.
There are three levels of difficulty: easy, medium, and hard, with different numbers of chances to guess depending on the level.
The user is prompted to choose a level and then play the game. They are given hints in the form of vowel or consonant for alphabet guesses.
If the user guesses the number or alphabet correctly within the given chances, they win the game.
If not, they lose. After each game, the user is asked whether they want to play again at the same level, choose another level, or go back to the menu.